requirements.txt
requests
python-dotenv
pandas
pyarrow
fastparquet
sqlalchemy
tenacity
pytest
